+++ app/src/main/java/com/aconno/sensorics/ui/readings/GenericReadingListFragment.kt
@@ -39,7 +39,7 @@ class GenericReadingListFragment : DaggerFragment() {
     override fun onPrepareOptionsMenu(menu: Menu) {
         super.onPrepareOptionsMenu(menu)
         activity?.menuInflater?.inflate(R.menu.menu_readings, menu)
-        menu.findItem(R.id.action_start_usecases_activity)?.isVisible = BuildConfig.FLAVOR == "dev"
+        menu.findItem(R.id.action_start_usecases_activity)?.isVisible = BuildConfig.DEBUG
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {